+++ package.json
@@ -181,6 +181,7 @@
         "semver-regex": "^2.0.0",
         "tslint": "^5.11.0",
         "tslint-eslint-rules": "^5.4.0",
+        "tslint-no-circular-imports": "^0.6.1",
         "typescript": "^3.0.1",
         "vscode": "^1.1.21",
         "vscode-nls-dev": "^3.2.2"
