+++ src/test/lambdaProvider.test.ts
@@ -3,9 +3,8 @@ import * as AWS from 'aws-sdk'
 import * as vscode from 'vscode'
 import { LambdaProvider } from '../lambda/lambdaProvider'
 import { RegionNode } from '../lambda/explorer/regionNode'
-import { ContextChangeEventsArgs } from '../shared/defaultAwsContext'
 import { AwsContextTreeCollection } from '../shared/awsContextTreeCollection'
-import { AwsContext } from '../shared/awsContext'
+import { AwsContext, ContextChangeEventsArgs } from '../shared/awsContext'
 import { RegionInfo } from '../shared/regions/regionInfo'
 import { RegionProvider } from '../shared/regions/regionProvider'
 import { ResourceFetcher } from '../shared/resourceFetcher'
